html/browsers/history/the-location-interface/location_hash.html

<!DOCTYPE HTML>
<html>
  <head>
    <title>location_hash</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <iframe id="srcdoc-iframe"
       srcdoc="<div style='height: 200vh'></div><div id='test'></div>"></iframe>
    <script>
//**This test assumes that assignments to location.hash will be synchronous - this is how all browsers implement it.
//The spec (as of 25 March 2011) disagrees.

    test(function () {
      window.history.pushState(1, document.title, '#x=1');
      var hash = location.hash;

      assert_equals(hash, "#x=1", "hash");

    }, "location hash");

    var t = async_test("Setting location.hash on srcdoc iframe");
    addEventListener("load", t.step_func_done(function() {
      var frameWin = document.getElementById("srcdoc-iframe").contentWindow;
      assert_equals(frameWin.location.href, "about:srcdoc");
      assert_equals(frameWin.scrollY, 0, "Should not have scrolled yet");
      frameWin.location.hash = "";
      assert_equals(frameWin.location.href, "about:srcdoc#", "Setting an empty hash should result in an empty fragment, not no fragment.");
      frameWin.location.hash = "test";
      assert_equals(frameWin.location.href, "about:srcdoc#test");
      assert_true(frameWin.scrollY > frameWin.innerHeight,
                  "Should have scrolled by more than one viewport height");
    }));

    test(function() {
      location.hash = "test";
      assert_equals(location.hash, "#test");
      location.hash = "";
    }, "Setting hash should automatically include hash character");

    test(function() {
      location.hash = "#not encoded";
      assert_equals(location.hash, "#not%20encoded");
      location.hash = "";
    }, "Setting hash should encode incompatible characters");

    test(function() {
      location.hash = "#already%20encoded";
      assert_equals(location.hash, "#already%20encoded");
      location.hash = "";
    }, "Setting hash to an already encoded value should not double encode it");

    test(function() {
      location.hash = "#mixed encoding%20here";
      assert_equals(location.hash, "#mixed%20encoding%20here");
      location.hash = "";
    }, "Setting hash which is partially encoded should only encode incompatible characters");
    </script>
  </body>
</html>
